commit latest updates
